nifty8 with new formula, tbt

+++ src/model_library/noise_models/fixed_noise.py
@@ -14,5 +14,5 @@ class StaticNoise(Model):
 
     def set_model(self):
         if self.inverse:
-            self._model = ift.makeOp(self.noise_cov**(-1))
-        self._model = ift.makeOp(self.noise_cov)
+            self._model = ift.makeOp(self.noise_cov**(-1),sampling_dtype=float)
+        self._model = ift.makeOp(self.noise_cov,sampling_dtype=float)
